requirements.txt
pandas
numpy
scikit-learn
xgboost
streamlit
shap
matplotlib
python-dotenv
openai
pyarrow